wizard/requirements.txt
# uDOS Wizard Pinned Dependencies (lockfile)
#
# Source of truth: pyproject.toml (+ Wizard runtime policy)
# This file is kept for compatibility with legacy pip installs and Docker builds.
#
# Preferred installation methods:
#   UV_PROJECT_ENVIRONMENT=.venv uv sync                     # Core/runtime baseline
#   UV_PROJECT_ENVIRONMENT=.venv uv sync --extra udos        # Standard ucode runtime
#   UV_PROJECT_ENVIRONMENT=.venv uv sync --extra udos-wizard --dev
#
# To regenerate this lockfile:
#   pip-compile pyproject.toml -o wizard/requirements.txt
#
# Health Check:
#   python3 -m core.services.self_healer

# =====================
# AI PROVIDERS (Wizard Server)
# =====================

# Gemini (Google AI)
google-generativeai>=0.3.0

# OpenAI
openai>=1.0.0

# Anthropic (Claude)
anthropic>=0.18.0

# =====================
# GOOGLE SERVICES (Wizard Server)
# =====================

# Gmail & Google Drive Integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# =====================
# IMAGE PROCESSING
# =====================

# Image Processing & Vectorization
Pillow>=10.0.0
vtracer>=0.6.0

# =====================
# CORE UTILITIES
# =====================

# Environment variable management
python-dotenv>=1.0.0

# Timezone detection
tzlocal>=4.0
pytz>=2023.3

# Interactive CLI with tab completion and arrow key support
# UBUNTU/DEBIAN USERS: Install system dependencies first:
#   sudo apt-get install libreadline-dev libncurses5-dev python3-dev
prompt_toolkit>=3.0.0

# JSON Schema validation (for checklist system)
jsonschema>=4.0.0

# YAML parsing (for prompt templates)
PyYAML>=6.0.0

# Encryption for private commands
cryptography>=41.0.0

# OS Keychain integration (macOS Keychain, Linux Secret Service)
keyring>=24.0.0

# HTTP requests for web extensions
requests>=2.31.0

# HTML parsing for website scraping
beautifulsoup4>=4.12.0

# System information
psutil>=5.9.0

# Password Hashing
bcrypt>=4.0.0

# =====================
# WEB SERVER (OPTIONAL - Wizard Server only)
# =====================
# NOTE: These are ONLY needed if running Wizard Server (port 8765)
# For Core TUI offline mode, you can skip these:
#   pip install -r requirements.txt --no-deps
#   pip install google-generativeai cryptography keyring python-dotenv ...
#
# OR use extras:
#   pip install -e .[wizard]     # Install with Wizard Server
#   pip install -e .[lite]       # Core only (no Wizard)

# FastAPI (Wizard Server) - REQUIRED
fastapi>=0.109.0
psycopg[binary]>=3.2.1
uvicorn>=0.27.0
python-multipart>=0.0.6

# Flask web framework - OPTIONAL
# flask>=2.0.0
# flask-cors>=5.0.0

# WebSocket support - OPTIONAL
# flask-socketio>=5.0.0
# simple-websocket>=1.0.0

# Async HTTP client - OPTIONAL
# aiohttp>=3.9.0
# hypercorn>=0.17.0

# Markdown rendering - OPTIONAL
# markdown2>=2.4.0
# pygments>=2.10.0

# Fuzzy string matching
thefuzz>=0.19.0

# =====================
# TUI
# =====================

# Rich terminal UI
rich>=13.0.0

# =====================
# TESTING
# =====================

pytest-asyncio>=0.23.0

# =====================
# NOTES
# =====================

# - All packages use minimum version constraints (>=) for flexibility
# - Python 3.8+ required (3.10+ recommended)
# - Node.js and npm required for Tauri apps (separate installation)
# - See INSTALLATION.md for detailed setup instructions
